+++ pyrit/chat/__init__.py
@@ -2,7 +2,6 @@
 # Licensed under the MIT license.
 
 from pyrit.chat.aml_online_endpoint_chat import AMLOnlineEndpointChat
-from pyrit.chat.azure_openai_chat import AzureOpenAIChat
 from pyrit.chat.hugging_face_chat import HuggingFaceChat
 
-__all__ = ["AzureOpenAIChat", "AMLOnlineEndpointChat", "HuggingFaceChat"]
+__all__ = ["AMLOnlineEndpointChat", "HuggingFaceChat"]
